apps/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pandas
scikit-learn
openai-whisper
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langchain-huggingface
langchain-ollama>=0.2.0
chromadb
sentence-transformers
pypdf
